% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_foro}
\alias{da_foro}
\title{Classficacao de foros e comarcas.}
\format{
um data frame com 7955 linhas and 5 variaveis:
\describe{
\item{id_justica}{numero de classificacao}
\item{id_tribunal}{classicacao pela da_justica}
\item{id_foro}{numero do respectivo foro}
\item{comarca}{comarca do respectivo foro}
\item{descricao}{foro}
}
}
\source{
\url{https://www.cnj.jus.br/programas-e-acoes/numeracao-unica/documentos/}
}
\usage{
da_foro
}
\description{
Um conjunto de dados que contem todos os foros e comarcas
classificado pelo orgao e tribunal.
}
\keyword{datasets}
