% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_foro_comarca}
\alias{da_foro_comarca}
\title{Classficacao de foros, comarcas e o nº do IBGE.}
\format{
um data frame com 7955 linhas and 8 variaveis:
\describe{
\item{id_justica}{numero de classificacao}
\item{id_tribunal}{classicacao pela da_justica}
\item{id_foro}{numero id do respectivo foro}
\item{comarca}{comarca do respectivo foro}
\item{descricao}{foro}
\item{sigla}{sigla do respectivo tribunal}
\item{uf}{estado do respectivo foto}
\item{ibge}{numero do ibge do respectivo foro}
}
}
\source{
\url{https://www.cnj.jus.br/programas-e-acoes/numeracao-unica/documentos/}
}
\usage{
da_foro_comarca
}
\description{
Um conjunto de dados que contem todos os foros e comarcas
classificado pelo orgao, tribunal e o nº do IBGE.
}
\keyword{datasets}
