% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_justica}
\alias{da_justica}
\title{Classificacao de orgao ou segmento do poder judiciario.}
\format{
um data frame com 9 linhas and 2 variaveis:
\describe{
\item{id_justica}{numero de classificacao}
\item{descricao}{orgao ou segmento judiciario}
}
}
\source{
\url{https://www.cnj.jus.br/programas-e-acoes/numeracao-unica/documentos/}
}
\usage{
da_justica
}
\description{
Um conjunto de dados que contem os orgaos e segmentos do poder
judiciario de 1:9.
}
\keyword{datasets}
