% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_tribunal}
\alias{da_tribunal}
\title{da_tribunal}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 106 rows and 5 columns.
}
\usage{
da_tribunal
}
\description{
da_tribunal
}
\keyword{datasets}
