% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_foro}
\alias{da_foro}
\title{da_foro}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7955 rows and 5 columns.
}
\usage{
da_foro
}
\description{
da_foro
}
\keyword{datasets}
