% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_justica}
\alias{da_justica}
\title{da_justica}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
da_justica
}
\description{
da_justica
}
\keyword{datasets}
