% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_tribunal}
\alias{da_tribunal}
\title{Levantamento dos Tribunais.}
\format{
um data frame com 106 linhas and 5 variaveis:
\describe{
\item{id_justica}{numero de classificacao}
\item{id_tribunal}{classicacao pela da_justica}
\item{descricao}{tribunal}
\item{sigla}{sigla do respectivo tribunal}
\item{uf}{estado do respectivo tribunal}
}
}
\source{
\url{https://www.cnj.jus.br/programas-e-acoes/numeracao-unica/documentos/}
}
\usage{
da_tribunal
}
\description{
Um conjunto de dados que contem os Tribunais Federais, Estaduais,
e do Distrito Federal e Territórios, Tribuais Regional eleitoral
e Tribunais Regionais do Trabalho.
}
\keyword{datasets}
